update: comment, likes, main, profiles, reports 코드 추가 및 userprofiles폴더 이름 변경

+++ src/likes/like.service.js
@@ -1,9 +1,14 @@
 const LikeRepository = require('./like.repository');
+const CustomError = require('../middlewares/error.middleware');
 
 class LikeService {
   likeRepository = new LikeRepository();
 
   clickLike = async (postId, userId) => {
+    // 게시글 존재 여부 확인
+    const postExists = await this.likeRepository.checkPostExists(postId);
+    if (!postExists) throw new CustomError('존재하지 않는 게시글입니다.', 404);
+
     const existLike = await this.likeRepository.existLike(postId, userId);
     if (existLike) {
       const removeLike = await this.likeRepository.removeLike(postId, userId);
